Move deprecations to a single file
This de-clutters the actual public API
